packages/backend/pyproject.toml
[project]
name = "findog-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.35.0",
  "pydantic>=2.11.0",
  "email-validator>=2.0.0",
  "sqlmodel>=0.0.22",
  "psycopg2-binary>=2.9.0",
  "pyjwt[crypto]>=2.9.0",
  "bcrypt>=4.2.0",
  "python-multipart>=0.0.12",
  "pydantic-settings>=2.6.0",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
